update collision detection code

+++ raytracing/hitlist.h
@@ -1,6 +1,7 @@
 #ifndef HITLIST_H
 #define HITLIST_H
 
+#include<cstdio>
 #include "hit.h"
 
 class hittable_list: public hittable {
@@ -18,7 +19,7 @@ bool hittable_list::hit(const ray& r, float t_min, float t_max,
 
     hit_record temp_rec;
     bool hit_anything = false;
-    double closest_so_far = t_max;
+    float closest_so_far = t_max;
     for (int i = 0; i < list_size; i++) {
         if (list[i]->hit(r, t_min, closest_so_far, temp_rec)) {
             hit_anything = true;
